requirements.txt
discord==1.7.3
google-cloud-vision==2.4.2
requests==2.26.0